string_mod added and some cleaning
